Format codes

+++ AndyCategory/NSString+Andy.m
@@ -77,7 +77,7 @@
 {
     const char *str = [self UTF8String];
     unsigned char result[CC_MD5_DIGEST_LENGTH];
-    CC_MD5(str, strlen(str), result);
+    CC_MD5(str, (unsigned int)strlen(str), result);
     
     NSMutableString *ret = [NSMutableString stringWithCapacity:CC_MD5_DIGEST_LENGTH * 2];
     for(int i = 0; i < CC_MD5_DIGEST_LENGTH; i++)
@@ -90,16 +90,18 @@
 /**
  *  sha1加密
  */
-- (NSString *)andy_sha1HexDigest {
+- (NSString *)andy_sha1HexDigest
+{
     const char *cstr = [self UTF8String];
     NSData *data = [NSData dataWithBytes:cstr length:self.length];
     
     uint8_t digest[CC_SHA1_DIGEST_LENGTH];
     
-    CC_SHA1(data.bytes, data.length, digest);
+    CC_SHA1(data.bytes, (unsigned int)data.length, digest);
     
     NSMutableString *ret = [NSMutableString stringWithCapacity:CC_SHA1_DIGEST_LENGTH * 2];
-    for(int i = 0; i < CC_SHA1_DIGEST_LENGTH; i++) {
+    for(int i = 0; i < CC_SHA1_DIGEST_LENGTH; i++)
+    {
         [ret appendFormat:@"%02x", digest[i]];
     }
     return ret;
@@ -202,7 +204,8 @@
 - (BOOL)andy_isValidUserIdCard
 {
     BOOL flag;
-    if (self.length <= 0) {
+    if (self.length <= 0)
+    {
         flag = NO;
         return flag;
     }
@@ -268,17 +271,22 @@
 }
 
 // 格式化手机号
-- (NSString *)andy_standardPhone {
+- (NSString *)andy_standardPhone
+{
     NSInteger position = 3;
     
     NSMutableString *tmp = [NSMutableString stringWithFormat:@"%@", self];
     [tmp replaceOccurrencesOfString:@" " withString:@"" options:NSCaseInsensitiveSearch range:NSMakeRange(0, tmp.length)];
     
-    while (1) {
-        if (tmp.length > position) {
+    while (1)
+    {
+        if (tmp.length > position)
+        {
             [tmp insertString:@" " atIndex:position];
             position += 5;
-        } else {
+        }
+        else
+        {
             break;
         }
     }
@@ -289,7 +297,8 @@
 // 格式化手机号
 - (NSString *)andy_standardTele
 {
-    if (self.length < 8) {
+    if (self.length < 8)
+    {
         return self;
     }
     
@@ -300,8 +309,10 @@
 }
 
 // 手机号****处理
-- (NSString *)andy_securitPhone {
-    if (self.length != 11) {
+- (NSString *)andy_securitPhone
+{
+    if (self.length != 11)
+    {
         return self;
     }
     return [self stringByReplacingCharactersInRange:NSMakeRange(3,4) withString:@"****"];
@@ -328,11 +339,13 @@
     NSData *result = [data andy_aes256_encrypt:key];
     
     //转换为2进制字符串
-    if (result && result.length > 0) {
+    if (result && result.length > 0)
+    {
         
         Byte *datas = (Byte*)[result bytes];
         NSMutableString *output = [NSMutableString stringWithCapacity:result.length * 2];
-        for(int i = 0; i < result.length; i++){
+        for(int i = 0; i < result.length; i++)
+        {
             [output appendFormat:@"%02x", datas[i]];
         }
         return output;
@@ -348,7 +361,8 @@
     unsigned char whole_byte;
     char byte_chars[3] = {'\0','\0','\0'};
     int i;
-    for (i=0; i < [self length] / 2; i++) {
+    for (i=0; i < [self length] / 2; i++)
+    {
         byte_chars[0] = [self characterAtIndex:i*2];
         byte_chars[1] = [self characterAtIndex:i*2+1];
         whole_byte = strtol(byte_chars, NULL, 16);
@@ -357,7 +371,8 @@
     
     //对数据进行解密
     NSData* result = [data andy_aes256_decrypt:key];
-    if (result && result.length > 0) {
+    if (result && result.length > 0)
+    {
         return [[NSString alloc] initWithData:result encoding:NSUTF8StringEncoding];
     }
     return nil;
@@ -369,7 +384,8 @@
 
 - (instancetype)andy_safe_initWithString:(NSString *)aString
 {
-    if (aString == nil || [aString isKindOfClass:[NSNull class]]) {
+    if (aString == nil || [aString isKindOfClass:[NSNull class]])
+    {
         return [self initWithString:@""];
     }
     
@@ -378,11 +394,13 @@
 
 - (NSString *)andy_safe_substringToIndex:(NSInteger)to
 {
-    if (to <= 0) {
+    if (to <= 0)
+    {
         return @"";
     }
     
-    if (to >= self.length) {
+    if (to >= self.length)
+    {
         return self;
     }
     
@@ -391,11 +409,13 @@
 
 - (NSString *)andy_safe_substringFromIndex:(NSInteger)from
 {
-    if (from <= 0) {
+    if (from <= 0)
+    {
         return self;
     }
     
-    if (from >= self.length) {
+    if (from >= self.length)
+    {
         return @"";
     }
